rename logsyntaxbuilder to objectlogsyntaxbuilder
