--- dspot/src/main/java/eu/stamp_project/dspot/assertiongenerator/assertiongenerator_components/methodreconstructor_components/observer_components/testwithloggenerator_components/logsyntaxbuilder_constructs/objectlog_components/FailToObserveException.java
@@ -1,9 +1,0 @@
-package eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.methodreconstructor_components.observer_components.testwithloggenerator_components.logsyntaxbuilder_constructs.objectlog_components;
-
-/**
- * Created by Benjamin DANGLOT
- * benjamin.danglot@inria.fr
- * on 05/04/18
- */
-public class FailToObserveException extends Exception {
-}
