+++ dspot/src/test/java/eu/stamp_project/test_framework/TestFrameworkTest.java
@@ -54,7 +54,7 @@ public class TestFrameworkTest extends AbstractTest {
         final String expectedToString = "public static junit.framework.Test suite() {" + AmplificationHelper.LINE_SEPARATOR +
                 "    return new junit.extensions.TestSetup(new junit.framework.TestSuite(fr.inria.helper.SecondClassJUnit3.class)) {" + AmplificationHelper.LINE_SEPARATOR +
                 "        protected void tearDown() throws java.lang.Exception {" + AmplificationHelper.LINE_SEPARATOR +
-                "            eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.methodreconstructor_components.observer_components.testwithloggenerator_components.logsyntaxbuilder_constructs.ObjectLog.save();" + AmplificationHelper.LINE_SEPARATOR +
+                "            eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.methodreconstructor_components.observer_components.testwithloggenerator_components.objectlogsyntaxbuilder_constructs.ObjectLog.save();" + AmplificationHelper.LINE_SEPARATOR +
                 "        }" + AmplificationHelper.LINE_SEPARATOR +
                 "    };" + AmplificationHelper.LINE_SEPARATOR +
                 "}";
